+++ .gitignore
@@ -7,3 +7,4 @@ MobileSdkRs/RustFramework*
 jna-*.jar
 *~
 kotlin/mobilesdkrs/src/main/java/com/spruceid/mobile/sdk/rs/mobile_sdk_rs.kt
+.DS_Store
